fix(zephyr): correct PSRAM Kconfig symbols + Zephyr SMH allocator API
The PSRAM Kconfig symbols were ESP-IDF names (CONFIG_SPIRAM,
CONFIG_SPIRAM_TRY_ALLOCATE_WIFI_LWIP, CONFIG_SPIRAM_BOOT_INIT) which
don't exist in Zephyr — Kconfig aborted with "attempt to assign the
value 'y' to the undefined symbol SPIRAM". Fixed to the correct Zephyr
symbols from soc/espressif/common/Kconfig.spiram:
  CONFIG_ESP_SPIRAM=y (master switch, selects SHARED_MULTI_HEAP)
  CONFIG_SPIRAM_MODE_OCT=y (OPI/ESP32-S3) or CONFIG_SPIRAM_MODE_QUAD=y

The Zephyr PSRAM allocator used ESP-IDF's heap_caps_malloc/
MALLOC_CAP_SPIRAM API, which doesn't exist in Zephyr. Zephyr uses the
shared multi-heap (SMH): the ESP32 SoC code registers PSRAM as an
SMH_REG_ATTR_EXTERNAL region at boot; allocation is via
shared_multi_heap_alloc(SMH_REG_ATTR_EXTERNAL, bytes). Updated
ui_psram_malloc and the psramFound shim accordingly.

Also fixed the long-standing -Wformat warning in the select/checkbox
auto-text codegen: snprintf's size param is size_t but the textFn
signature uses uint8_t — cast to static_cast<size_t>(size). This
warning was pre-existing but only blocked the build when PSRAM's
Kconfig change triggered a full reconfigure.

Verified: demo-st compiles --compile clean with psram:'opi' enabled
(ELF produced, Kconfig accepted, BOARD_HAS_PSRAM defined, SMH allocator
present). Re-commented in the demo config (devkitc has no PSRAM).
18/18 kconfig tests green (updated for the correct symbol names).

+++ packages/cuttlefish/src/ir/ui-element-auto-wire.ts
@@ -112,7 +112,7 @@ function autoWireNode(treeName: string, node: AutoWireNode, nodeIndex: number): 
         nodeIndex,
         property: "text",
         fnName: `__ui_bindtext_${nodeIndex}`,
-        cppBody: `snprintf(buf, size, "%s", ${sig});`,
+        cppBody: `snprintf(buf, static_cast<size_t>(size), "%s", ${sig});`,
       });
       // Write: keyboard commit → signal.set(text).
       recordInputBinding({
@@ -191,12 +191,14 @@ function autoWireNode(treeName: string, node: AutoWireNode, nodeIndex: number): 
       callbackBody: `__ui_nodes[${nodeIndex}].value = (__ui_nodes[${nodeIndex}].value + 1) % ${count};`,
     });
 
-    // Auto-bind text to show the current option via snprintf if/else chain
+    // Auto-bind text to show the current option via snprintf if/else chain.
+    // Cast size to size_t to satisfy -Wformat (snprintf's n param is size_t;
+    // the textFn signature uses uint8_t).
     const branches = options.map((opt, i) => {
-      if (i === 0) return `if (__ui_nodes[${nodeIndex}].value == 0) { snprintf(buf, size, "%s", "${opt}"); }`;
-      return `else if (__ui_nodes[${nodeIndex}].value == ${i}) { snprintf(buf, size, "%s", "${opt}"); }`;
+      if (i === 0) return `if (__ui_nodes[${nodeIndex}].value == 0) { snprintf(buf, static_cast<size_t>(size), "%s", "${opt}"); }`;
+      return `else if (__ui_nodes[${nodeIndex}].value == ${i}) { snprintf(buf, static_cast<size_t>(size), "%s", "${opt}"); }`;
     }).join(" ");
-    const elseBranch = `else { snprintf(buf, size, "%s", "${options[0] || ""}"); }`;
+    const elseBranch = `else { snprintf(buf, static_cast<size_t>(size), "%s", "${options[0] || ""}"); }`;
 
     recordBinding({
       nodeIndex,
